requirements.txt
Flask==1.1.2
Flask_Cors==3.0.9
Flask_RESTful==0.3.9
keras==2.6.0
matplotlib==3.5.1
numpy==1.19.5
pandas==1.4.1
Pillow==9.1.0
scikit_learn==1.0.2
scipy==1.8.0
tensorflow==2.5.3
